+++ HISTORY.rst
@@ -4,6 +4,10 @@ Changelog
 0.12 - Unreleased
 -----------------
 
+* Support ``proxycommand`` option for plain instances. This also caused a
+  change in the ``init_ssh_key`` API for plugins.
+  [fschulze]
+
 * Support ``ProxyCommand`` from ``~/.ssh/config`` for plain instances.
   Requires Fabric 1.5.0 and Paramiko 1.9.0 or newer.
   [fschulze]